Create AzureKeyVaultPowerShell
This script uses 4 pieces of information to retrieve a secret from the Azure Key Vault. First it uses the TenantId which is also the same as the subscription Id, An Azure Application Id, and a thumbprint of a locally installed certificate that is associated with the Azure Application Id which has been granted permission to access the key vault.